会话列表、离线消息、群聊等功能实现

+++ server/internal/message/model.go
@@ -11,7 +11,7 @@ type Message struct {
 	FromUid   int64     `gorm:"not null" json:"from_uid"`
 	ToUid     int64     `gorm:"not null" json:"to_uid"`
 	Content   string    `gorm:"type:text" json:"content"`
-	MsgType   int      `gorm:"type:tinyint;not null" json:"msg_type"`
+	MsgType   int8      `gorm:"type:tinyint;not null" json:"msg_type"`
 	IsRevoke  int8      `gorm:"type:tinyint;default:0" json:"is_revoke"`
 	CreatedAt time.Time `json:"created_at"`
 }
